requirements.txt
Django==5.0.7
django-allauth==0.63.6
django-rest-auth==0.9.5
dj-rest-auth==6.0.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
requests==2.32.3